backend/requirements.txt
fastapi==0.115.12
pydantic-settings==2.8.1
uvicorn==0.34.1
python-dotenv
google-genai
